% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dog_years_function.R
\name{dog_years_function}
\alias{dog_years_function}
\title{A Dog Years Function
This function allows you to know your age in dog years.}
\usage{
dog_years_function(age)
}
\arguments{
\item{age}{What is your age?}
}
\description{
A Dog Years Function
This function allows you to know your age in dog years.
}
\examples{
dog_years_function(12)
}
\keyword{dog}
\keyword{years}
